refactor: gateway
